Add support cascade replication
pg_current_xlog_location() returns the current transaction log write
location but it only works if postgresql is not in recovery.

pg_last_xlog_receive_location() get last transaction log location
received and synced to disk by streaming replication. While the
streaming is in progress this will increase monotonically. This function
can be called on any replica meanwhile is in recovery.

We use the latter to calculate the different in bytes bettween the last
transaction log location recived by the replica and the replay_location
reported by the cascade replica.